Allow the user to draw with color on the canvas
To allow the user to use multiple colors when drawing, inline style was
applied to the canvas grid instead CSS classes. The color input was
modified to allow the user to change the color just pressing its parent
button. unselectButtons() was created because the last version only
worked for two buttons. getRandomColor() returns a random color, it will
be used with the rainbow mode not yet functional